Bug fixed
An important bgug fixed.

SIM +eps was not consdiered for all cases, allowing devision to zero for rating prediction and creating of Nan entries. This was fixed.

Also, sparse indexing was changed to regular indexing for efficiency matters. It helped improvement of the rating prediction by 1.5 times.

The code also improved and can be considered as the first final version.